Improve efficiency of _hunspellFindMisspelledWords
`_findWordWithinString` will now return the character number of the word character that follows the found string. `_hunspellFindMisspelledWords` will use this value when determing the next character to scan after finding a word in a string.

In the following example the old parser would find `test` and then start scanning on `”`. Because `”` isn’t a word character `_findWordWithinString` would look backwards. Now it will always search forwards because it will start on `s` in `string`.

```
This is a “test” string.
```